requirements.txt
# Core AI/LLM
anthropic>=0.40.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Vector DB and embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Data sources
yfinance>=0.2.40
requests>=2.32.0
feedparser>=6.0.11

# Caching
redis>=5.0.0

# Web UI
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0

# Configuration and utilities
pyyaml>=6.0.2
python-dotenv>=1.0.0
boto3>=1.35.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# Text processing
tiktoken>=0.8.0
